Migrate code to support ckan 2.9 and python3.
