--- ckanext/google_tag_manager/plugin.py
@@ -1,24 +1,0 @@
-import ckan.plugins as plugins
-import ckan.plugins.toolkit as toolkit
-import pylons.config as config
-
-
-def get_gtm_code():
-    # To get Google Tag Manager Code
-    gtm_code = config.get('ckan.google_tag_manager.gtm_container_id', False)
-    return str(gtm_code)
-
-class Google_Tag_ManagerPlugin(plugins.SingletonPlugin):
-    plugins.implements(plugins.IConfigurer)
-
-    # IConfigurer
-
-    def update_config(self, config_):
-        toolkit.add_template_directory(config_, 'templates')
-        toolkit.add_public_directory(config_, 'public')
-        toolkit.add_resource('fanstatic', 'google_tag_manager')
-    
-    # Declare that this plugin will implement ITemplateHelpers.
-    plugins.implements(plugins.ITemplateHelpers)
-    def get_helpers(self):
-        return {'get_gtm_container_id' : get_gtm_code}
